Moved command modules to a separate package
