+++ pypixelart/command/command.py
@@ -3,7 +3,8 @@ from typing import Protocol
 
 class Command(Protocol):
     """
-    Command class implemented by PyPixelArt actions, following the Command Design Pattern
+    Command class implemented by PyPixelArt actions that make changes to the image,
+    following the Command Design Pattern
     """
 
     def execute(self) -> None:
